+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The `Unreleased` section name is replaced by the expected version of next releas
 - `eqx dump`'s `-J` switch now turns off JSON rendering
 - `eqx -C dump` now renders Store `.Information` logs
 - Samples consistently use `module Fold`, `(Events.ForX streamId)`, removed `(|Stream|)`  [#174](https://github.com/jet/equinox/pull/174) 
+- Targeted `FsCodec` v `2.0.0-rc1`
 
 ### Removed
 ### Fixed